refactor: HttpRequestParser 분리
