+++ tomcat/src/main/java/org/apache/coyote/http/request/HttpRequest.java
@@ -1,6 +1,5 @@
 package org.apache.coyote.http.request;
 
-import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
@@ -16,11 +15,11 @@ public class HttpRequest {
     private final HttpCookie cookies;
     private final SessionManager sessionManager;
 
-    public HttpRequest(BufferedReader reader, SessionManager sessionManager) throws IOException {
-        this.requestLine = parseRequestLine(reader);
-        this.headers = Map.copyOf(parseHeaders(reader));
-        this.body = parseBody(reader, headers);
-        this.cookies = HttpCookie.parse(headers.get("Cookie"));
+    public HttpRequest(RequestLine requestLine, Map<String, String> headers, String body, HttpCookie cookies, SessionManager sessionManager) {
+        this.requestLine = requestLine;
+        this.headers = Map.copyOf(headers);
+        this.body = body;
+        this.cookies = cookies;
         this.sessionManager = sessionManager;
     }
 
@@ -40,48 +39,6 @@ public class HttpRequest {
         return Map.copyOf(params);
     }
 
-    private Map<String, String> parseHeaders(BufferedReader reader) throws IOException {
-        final var headers = new HashMap<String, String>();
-        String line;
-
-        while ((line = reader.readLine()) != null && !line.isEmpty()) {
-            String[] headerParts = line.split(":", 2);
-            if (headerParts.length == 2) {
-                headers.put(headerParts[0].trim(), headerParts[1].trim());
-            }
-        }
-
-        return headers;
-    }
-
-    private RequestLine parseRequestLine(BufferedReader reader) throws IOException {
-        final var line = reader.readLine();
-        if (line == null) {
-            throw new IllegalArgumentException("요청 라인이 없음");
-        }
-
-        final var parts = line.split(" ");
-        if (parts.length != 3) {
-            throw new IllegalArgumentException("잘못된 요청 라인: " + line);
-        }
-
-        return new RequestLine(parts[0], parts[1], parts[2]);
-    }
-
-    private String parseBody(BufferedReader reader, Map<String, String> headers) throws IOException {
-        if (!headers.containsKey("Content-Length")) {
-            return "";
-        }
-
-        try {
-            int contentLength = Integer.parseInt(headers.get("Content-Length"));
-            char[] buffer = new char[contentLength];
-            reader.read(buffer, 0, contentLength);
-            return new String(buffer);
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("잘못된 Content-Length: " + headers.get("Content-Length"));
-        }
-    }
 
     public Session getSession(boolean create) {
         String sessionId = cookies.getJSessionId();
