+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -5,10 +5,10 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.Socket;
+import org.apache.catalina.session.SessionManager;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http.handler.RequestHandler;
-import org.apache.coyote.http.request.HttpRequest;
-import org.apache.catalina.session.SessionManager;
+import org.apache.coyote.http.request.HttpRequestParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -17,12 +17,12 @@ public class Http11Processor implements Runnable, Processor {
     private static final Logger log = LoggerFactory.getLogger(Http11Processor.class);
     private final Socket connection;
     private final RequestHandler requestHandler;
-    private final SessionManager sessionManager;
+    private final HttpRequestParser requestParser;
 
     public Http11Processor(final Socket connection, final SessionManager sessionManager) {
         this.connection = connection;
-        this.sessionManager = sessionManager;
         this.requestHandler = new RequestHandler();
+        this.requestParser = new HttpRequestParser(sessionManager);
     }
 
     @Override
@@ -37,7 +37,7 @@ public class Http11Processor implements Runnable, Processor {
                 final var outputStream = connection.getOutputStream()) {
 
             final var bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
-            final var request = new HttpRequest(bufferedReader, sessionManager);
+            final var request = requestParser.parse(bufferedReader);
             final var response = requestHandler.handleRequest(request);
             response.writeTo(outputStream);
